indexer/requirements.txt
elasticsearch>=8.0.0
langchain
langchain-community
httpx
python-dotenv
